fix: add TaskNotificationManager
+ move notification work starting to application
+ worker -> coroutineWorker
