+++ app/src/main/java/com/example/plantsapp/presentation/PlantApplication.kt
@@ -1,11 +1,19 @@
 package com.example.plantsapp.presentation
 
 import android.app.Application
+import android.content.Context
 import androidx.hilt.work.HiltWorkerFactory
 import androidx.work.Configuration
+import androidx.work.ExistingPeriodicWorkPolicy
+import androidx.work.PeriodicWorkRequestBuilder
+import androidx.work.WorkManager
 import com.example.plantsapp.BuildConfig
+import com.example.plantsapp.presentation.ui.notification.NotificationWorker
+import com.example.plantsapp.presentation.ui.utils.calculateDelay
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
+import java.util.*
+import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 @HiltAndroidApp
@@ -20,10 +28,35 @@ class PlantApplication : Application(), Configuration.Provider {
         if (BuildConfig.DEBUG) {
             Timber.plant(Timber.DebugTree())
         }
+
+        startNotificationWorker(
+            context = applicationContext,
+            startDate = Calendar.getInstance()
+        )
     }
 
     override fun getWorkManagerConfiguration(): Configuration =
         Configuration.Builder()
             .setWorkerFactory(workerFactory)
             .build()
+
+    private fun startNotificationWorker(context: Context, startDate: Calendar) {
+        val notificationRequest =
+            PeriodicWorkRequestBuilder<NotificationWorker>(1, TimeUnit.DAYS)
+                /*.setInitialDelay(
+                    startDate.calculateDelay(nextDayHour = HOUR_OF_WORK_STARTING),
+                    TimeUnit.MILLISECONDS
+                )*/
+                .build()
+
+        WorkManager.getInstance(context).enqueueUniquePeriodicWork(
+            NotificationWorker.NOTIFICATION_WORK_NAME,
+            ExistingPeriodicWorkPolicy.REPLACE,
+            notificationRequest
+        )
+    }
+
+    companion object {
+        private const val HOUR_OF_WORK_STARTING = 10
+    }
 }
